requirements.txt
robotframework==5.0.1
PyYAML==6.0
Appium-Python-Client==2.10.1
robotframework-seleniumlibrary==6.0.0
robotframework-appiumlibrary==2.0